refactor: extract `IndexerConfig` to own mod
